ci: add support to CentOS
Currently the CI for TDX runs on CentOS Stream 8. The ansible playbooks
where changed to install/configure on that distro as well.

Added a new vagrant VM configuration 'tests-e2e-centosstream8', which is
not fully functional because it creates a disk with only 10GB. Even so, if
you want to run it locally (it should fail to install the operator due
the lack of disk space):

$ vagrant up tests-e2e-centosstream8

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>